fix clearWebRequest button
